requirements.txt
PyMuPDF
langchain
openai
dotenv
weaviate-client
python-docx
streamlit